:recycle: BTS-300/refactor: 타입 파일 정리
