+++ src/features/auth/services/api.ts
@@ -3,7 +3,7 @@ import {
   LoginBody,
   SignUpBody,
   VerifyCodeBody,
-} from '@/features/auth/type';
+} from '@/features/auth/types';
 import type { Member } from '@/features/member/model/types';
 import { authApi, publicApi } from '@/lib/http/api';
